+++ forui/lib/src/widgets/checkbox.dart
@@ -210,6 +210,7 @@ class FCheckboxStyle with Diagnosticable {
         );
 
   /// The [FLabel]'s style.
+  // ignore: diagnostic_describe_all_properties
   FLabelStyle get labelStyle => (
         layout: labelLayoutStyle,
         state: FLabelStateStyles(
@@ -247,8 +248,7 @@ class FCheckboxStyle with Diagnosticable {
       ..add(DiagnosticsProperty('labelLayoutStyle', labelLayoutStyle))
       ..add(DiagnosticsProperty('enabledStyle', enabledStyle))
       ..add(DiagnosticsProperty('disabledStyle', disabledStyle))
-      ..add(DiagnosticsProperty('errorStyle', errorStyle))
-      ..add(DiagnosticsProperty('labelStyle', labelStyle));
+      ..add(DiagnosticsProperty('errorStyle', errorStyle));
   }
 
   @override