+++ forui/lib/src/widgets/switch.dart
@@ -221,6 +221,7 @@ final class FSwitchStyle with Diagnosticable {
         );
 
   /// The [FLabel]'s style.
+  // ignore: diagnostic_describe_all_properties
   FLabelStyle get labelStyle => (
         layout: labelLayoutStyle,
         state: FLabelStateStyles(
@@ -255,8 +256,7 @@ final class FSwitchStyle with Diagnosticable {
       ..add(DiagnosticsProperty('labelLayoutStyle', labelLayoutStyle))
       ..add(DiagnosticsProperty('enabledStyle', enabledStyle))
       ..add(DiagnosticsProperty('disabledStyle', disabledStyle))
-      ..add(DiagnosticsProperty('errorStyle', errorStyle))
-      ..add(DiagnosticsProperty('labelStyle', labelStyle));
+      ..add(DiagnosticsProperty('errorStyle', errorStyle));
   }
 
   @override